Fix for 3.5
